split race and norace tests
