+++ beacon-chain/blockchain/BUILD.bazel
@@ -38,8 +38,16 @@ go_library(
     ],
 )
 
-go_test(
+test_suite(
     name = "go_default_test",
+    tests = [
+        ":go_raceoff_test",
+        ":go_raceon_test",
+    ],
+)
+
+go_test(
+    name = "go_raceoff_default_test",
     size = "medium",
     srcs = [
         "chain_info_test.go",
@@ -47,8 +55,42 @@ go_test(
         "receive_block_test.go",
         "service_test.go",
     ],
+    embed = [":go_default_library"],    
+    deps = [
+        "//beacon-chain/cache/depositcache:go_default_library",
+        "//beacon-chain/core/blocks:go_default_library",
+        "//beacon-chain/core/helpers:go_default_library",
+        "//beacon-chain/core/state:go_default_library",
+        "//beacon-chain/db:go_default_library",
+        "//beacon-chain/db/testing:go_default_library",
+        "//beacon-chain/p2p:go_default_library",
+        "//beacon-chain/powchain:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+        "//proto/eth/v1alpha1:go_default_library",
+        "//shared/bytesutil:go_default_library",
+        "//shared/event:go_default_library",
+        "//shared/params:go_default_library",
+        "//shared/testutil:go_default_library",
+        "@com_github_ethereum_go_ethereum//:go_default_library",
+        "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
+        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+        "@org_golang_x_net//context:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_raceon_test",
+    srcs = [
+        "service_norace_test.go",
+        "chain_info_norace_test",
+    ],
     embed = [":go_default_library"],
     race = "on",
+    tags = ["race_on"],
     deps = [
         "//beacon-chain/cache/depositcache:go_default_library",
         "//beacon-chain/core/blocks:go_default_library",
